webapi/tct-contact-tizen-tests/contact/AddressBook_addBatch_two_oneIsNull.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Cici Li <cici.x.li@intel.com>
        Xu, Yananx <yananx.xu@intel.com>

-->
<html>
<head>
<title>AddressBook_addBatch_two_oneIsNull</title>
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">
//==== TEST: AddressBook_addBatch_two_oneIsNull
//==== LABEL Check if method AddressBook::addBatch() can add two times the same contact
//==== PRIORITY P2
//==== SPEC Tizen Web API:Social:Contact:AddressBook:addBatch M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/contact.html
//==== TEST_CRITERIA MOA
var t = async_test(document.title), defaultAddressBook, contactA, contactB, contactsArray, batchAddedSuccessCB,
    batchAddFailedCB, i;
t.step(function () {
    contactA = new tizen.Contact({
        name: new tizen.ContactName({
            firstName: "Jeffrey",
            lastName: "Hyman"
        })
    });
    contactB = new tizen.Contact(null);

    contactsArray = [contactA, contactB];
    defaultAddressBook = tizen.contact.getDefaultAddressBook();
    batchAddedSuccessCB = t.step_func(function (contacts) {
        assert_equals(contacts.length, 2, "add two contacts here");
        for (i = 0; i < contacts.length; i++) {
            defaultAddressBook.remove(contacts[i].id);
        }
        t.done();
    });

    batchAddFailedCB = t.step_func(function (error) {
        assert_unreached("Invalid errorCallback invoked: " + error.name + ": " + error.message);
    });
    defaultAddressBook.addBatch(contactsArray, batchAddedSuccessCB, batchAddFailedCB);
});

</script>
</body>
</html>
